feat: Refactor imports to use absolute paths and restructure collections for authentication, posts, roles, and users

+++ src/modules/services/post.service.ts
@@ -1,16 +1,25 @@
 import { PostCreate, PostResponse, PostUpdate } from "@models/post.model";
 import { IPostRepository } from "@repositories/post.repository";
-import { PostMapper } from "modules/mappers/post.mapper";
+import { PostMapper } from "@mappers/post.mapper";
+import { RoleResponse } from "@models/role.model";
+import { NotFoundError } from "@errors/not-found-error";
 
 interface IPostService {
     findPosts(): Promise<PostResponse[]>;
     findPostById(id: string): Promise<PostResponse | null>;
+    // todo: Should I send the userId ?
     createPost(newPost: PostCreate): Promise<PostResponse | null>;
     updatePost(
         id: string,
         partialPost: PostUpdate,
+        userRoles: RoleResponse[],
+        loggedInUserId: string,
     ): Promise<PostResponse | null>;
-    deletePost(id: string): Promise<boolean>;
+    deletePost(
+        id: string,
+        userRoles: RoleResponse[],
+        loggedInUserId: string,
+    ): Promise<boolean>;
     //
     existsByIdAndUserId(id: string, userId: string): Promise<boolean>;
 }
@@ -29,6 +38,10 @@ class PostService implements IPostService {
 
     async findPostById(id: string): Promise<PostResponse | null> {
         const existingPost = await this.repository.findById(id);
+
+        if (!existingPost)
+            throw new NotFoundError(`Post with this id ${id} not found.`);
+
         return existingPost ? PostMapper.toDTO(existingPost) : null;
     }
 
@@ -40,12 +53,62 @@ class PostService implements IPostService {
     async updatePost(
         id: string,
         partialPost: PostUpdate,
+        userRoles: RoleResponse[],
+        loggedInUserId: string,
     ): Promise<PostResponse | null> {
+        // Check is the post with if exists.
+        const existingPost = await this.repository.findById(id);
+        if (!existingPost)
+            throw new NotFoundError(`Post with this id ${id} not found.`);
+
+        // Check if the standard user
+        const isStandardUser = userRoles.some(({ name }) => name === "user");
+        // Enforce ownership
+        if (isStandardUser) {
+            const existingPostWithUser = await this.repository.existsBy({
+                _id: id,
+                userId: loggedInUserId,
+            });
+
+            if (!existingPostWithUser)
+                throw new NotFoundError(
+                    `Post not found or you do not have permission to update it.`,
+                );
+        }
+
+        PostMapper.updateEntity(existingPost, partialPost);
+
         const postUpdated = await this.repository.update(id, partialPost);
+
         return postUpdated ? PostMapper.toDTO(postUpdated) : null;
     }
 
-    async deletePost(id: string): Promise<boolean> {
+    async deletePost(
+        id: string,
+        userRoles: RoleResponse[],
+        loggedInUserId: string,
+    ): Promise<boolean> {
+        // Check is the post with if exists.
+        const existingPost = await this.repository.existsBy({ _id: id });
+        
+        if (!existingPost)
+            throw new NotFoundError(`Post with this id ${id} not found.`);
+
+        // Check if the standard user
+        const isStandardUser = userRoles.some(({ name }) => name === "user");
+        // Enforce ownership
+        if (isStandardUser) {
+            const existingPostWithUser = await this.repository.existsBy({
+                _id: id,
+                userId: loggedInUserId,
+            });
+
+            if (!existingPostWithUser)
+                throw new NotFoundError(
+                    `Post not found or you do not have permission to update it.`,
+                );
+        }
+
         return await this.repository.delete(id);
     }
